+++ Packages/com.unity.inputsystem/InputSystem/Devices/Precompiled/FastMouse.partial.cs
@@ -9,8 +9,8 @@ namespace UnityEngine.InputSystem
             // Changing these separately seems to not result in much of a difference
             // compared to just doing an InputState.Change with a complete MouseState.
             double now = InputRuntime.s_Instance.currentTime;
-            InputState.Change(delta, Vector2.zero, InputState.currentUpdateType, time: now);
-            InputState.Change(scroll, Vector2.zero, InputState.currentUpdateType, time: now);
+            InputState.Change(delta, Vector2.zero, now, InputState.currentUpdateType);
+            InputState.Change(scroll, Vector2.zero, now, InputState.currentUpdateType);
         }
 
         // For FastMouse, we know that our layout is MouseState so we can just go directly