+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -11,7 +11,7 @@ however, it has to be formatted properly to pass verification tests.
 ## [Unreleased] - yyyy-mm-dd
 - Fixed Composite binding isn't triggered after ResetDevice() called during Action handler [ISXB-746](https://issuetracker.unity3d.com/product/unity/issues/guid/ISXB-746)
 - Fixed resource designation for "d_InputControl" icon to address CI failure
-- Changed FastMouse.OnNextUpdate() to pass the current time as a parameter to InputState.Change() improving performance (case ISX-1414)
+- Changed FastMouse.OnNextUpdate() to pass the current time as a parameter to InputState.Change() avoiding a redundant scripting to native call to query the time (case ISX-1414)
 
 ## [1.8.2] - 2024-04-29